+++ src/main/java/de/fernunihagen/dna/scalephant/performance/osm/util/GeometricalStructure.java
@@ -17,11 +17,6 @@
  *******************************************************************************/
 package de.fernunihagen.dna.scalephant.performance.osm.util;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
@@ -36,7 +31,7 @@ public class GeometricalStructure implements Serializable {
 	private static final long serialVersionUID = -2119998223400835002L;
 	
 	/**
-	 * The ID of the sttructure
+	 * The ID of the structure
 	 */
 	protected final long id;
 	
@@ -89,33 +84,6 @@ public class GeometricalStructure implements Serializable {
 		return new BoundingBox((float) minX, (float) maxX, (float) minY, (float) maxY);
 	}
 
-	/**
-	 * Convert object to byte array
-	 * @return
-	 * @throws IOException 
-	 */
-	public byte[] toByteArray() throws IOException {
-		final ByteArrayOutputStream bOutputStream = new ByteArrayOutputStream();
-		final ObjectOutputStream objectOutputStream = new ObjectOutputStream(bOutputStream);
-
-		objectOutputStream.writeObject(this);
-		
-		bOutputStream.close();
-		return bOutputStream.toByteArray();
-	}
-	
-	/**
-	 * Construct object from byte array
-	 * @param bytes
-	 * @return
-	 * @throws IOException 
-	 * @throws ClassNotFoundException 
-	 */
-	public GeometricalStructure loadFromByteArray(final byte[] bytes) throws IOException, ClassNotFoundException {
-		final ByteArrayInputStream bInputStream = new ByteArrayInputStream(bytes);
-		final ObjectInputStream oInputStream = new ObjectInputStream(bInputStream);
-		return (GeometricalStructure) oInputStream.readObject();
-	}
 
 	public long getId() {
 		return id;