+++ changelog.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ### Version 0.2.0 (Beta-2) - TBA
+- New Feature: Java 8 is now required to build the project
 - Improvement: Added missing copyright header to all source files
+- Improvement: OSM import is now executed on a disk backed set, so huge imports (> memory) can be performed
 
 ### Version 0.2.0 (Beta-1) - 17.11.2016
 - New feature: Added a selftest